+++ src/main/java/calculator/Validator.java
@@ -1,13 +1,13 @@
 package calculator;
 
 public class Validator {
-    public static int validate(final String str) {
+    public static long validate(final String str) {
         if (str.isEmpty()) return 0;
 
         if (isNotDigit(str))
             throw new IllegalArgumentException("[ERROR] 지정되지 않은 구분자입니다.");
 
-        int num = Integer.parseInt(str);
+        long num = Long.parseLong(str);
         if (isNegative(num))
             throw new IllegalArgumentException("[ERROR] 음수는 입력할 수 없습니다.");
 
@@ -21,7 +21,7 @@ public class Validator {
         return false;
     }
 
-    private static boolean isNegative(int num) {
+    private static boolean isNegative(long num) {
         return num < 0;
     }
 }
